Move index page contents to component page
