+++ src/Client/Web/Playground/Bit.Client.Web.BlazorUI.Playground/Web/Pages/Component.razor.css
@@ -1,0 +1,3 @@
+.text-field-container {
+    width: 30%;
+}
