+++ src/Client/Web/Playground/Bit.Client.Web.BlazorUI.Playground/Web/Pages/Component.razor.scss
@@ -1,1 +1,3 @@
-﻿
+﻿.text-field-container {
+    width: 30%
+}
